♻️(recrutement-presentation) isolate RecrutementKanbanView and RecrutementListeView to recrutement_detail.py for more clarity
