+++ src/web/presentation/recruteur/urls.py
@@ -9,6 +9,10 @@ from presentation.recruteur.views.organismes import (
     InitEtapesRecrutementOrganismeView,
     OrganismeView,
 )
+from presentation.recruteur.views.recrutement_detail import (
+    RecrutementKanbanView,
+    RecrutementListeView,
+)
 from presentation.recruteur.views.recrutement_listes import (
     RecrutementsActifsView,
     RecrutementsArchivesView,
@@ -17,8 +21,6 @@ from presentation.recruteur.views.recrutements import (
     InitRecrutementEtapeView,
     RecrutementCandidaturesEtapeView,
     RecrutementEtapeView,
-    RecrutementKanbanView,
-    RecrutementListeView,
 )
 
 app_name = "recruteur"
